Numeric.Backprop.Iso

Description

A poor substitute for the Iso module in lens, providing the Iso type synonym and some sample useful Isos for usage with backprop, without incuring a lens dependency.

If you also import lens, you should only use this module for the Isos it exports, and not import the redefined Iso type synonym or 're'/'iso'/'review'.

Isomorphisms

type Iso s t a b = forall p f. (Profunctor p, Functor f) => p a (f b) -> p s (f t)

A family of isomorphisms. See Iso'.

type Iso' s a = Iso s s a a

An Iso' s a encodes an isomorphism between an s and an a. It basically lets you go from s -> a and back (from a -> s) while preserving structure. You can basically imagine an Iso' s a to be an (s -> a, a -> s) tuple.

You can get the "forward" direction of an Iso' with view:

view :: Iso'' s a -> (s -> a)

And the "backwards" direction with review:

review :: Iso'' s a -> (a -> s)

You can construct an Iso' using iso, giving the forward and backwards functions:

>>> myIso :: Iso' (Identity a) a
    myIso = iso runIdentity Identity
>>> view myIso (Identity "hello")
"hello"
>>> review myIso "hello"
Identity "hello"

One powerful thing about Iso's is that they're composable using .:

(.) :: Iso' c b -> Iso' b a -> Iso' c a

This is basically provided here so that this package doesn't incurr a lens dependecy, but if you already depend on lens, you should use the version from Control.Lens.Iso instead.

Construction and usage

iso :: (s -> a) -> (b -> t) -> Iso s t a b

Construct an Iso by giving the "forward" and "backward" direction functions:

>>> myIso :: Iso' (Identity a) a
    myIso = iso runIdentity Identity
>>> view myIso (Identity "hello")
"hello"
>>> review myIso "hello"
Identity "hello"

This is basically provided here so that this package doesn't incurr a lens dependecy, but if you already depend on lens, you should use the version from Control.Lens.Iso instead.

re :: Iso' s a -> Iso' a s

Reverse an Iso'. The forward function becomes the backwards function, and the backwards function becomes the forward function.

This is basically provided here so that this package doesn't incurr a lens dependecy, but if you already depend on lens, you should use the version from Control.Lens.Review instead.

review :: Iso s t a b -> b -> t

Get the "reverse" direction function from an Iso.

This is basically provided here so that this package doesn't incurr a lens dependecy, but if you already depend on lens, you should use the version from Control.Lens.Review instead.

view :: Getting a s a -> s -> a

view is a synonym for (^.):

>>> view _1 (1, 2)
1

The reason it's not in Lens.Micro is that view in lens has a more general signature:

view :: MonadReader s m => Getting a s a -> m a

So, you would be able to use this view with functions, but not in various reader monads. For most people this shouldn't be an issue; if it is for you, use view from microlens-mtl.

Useful Isos

coerced :: Coercible s a => Iso' s a

A useful Iso between two types with the same runtime representation.

gTuple :: (Generic a, Code a ~ '[as]) => Iso' a (Tuple as)

An Iso between a type that is a product type, and a tuple that contains all of its components. Uses Generics.SOP and the Generic typeclass.

>>> import qualified Generics.SOP as SOP
>>> data Foo = A Int Bool      deriving Generic
>>> instance SOP.Generic Foo
>>> 'view' 'gTuple' (A 10 True)
10 ::< True ::< Ø
>>> 'review' 'gTuple' (15 ::< False ::< Ø)
A 15 False

gSOP :: Generic a => Iso' a (Sum Tuple (Code a))

An Iso between a sum type whose constructors are products, and a sum (Sum) of products (Tuple). Uses Generics.SOP and the Generic typeclass.

>>> import qualified Generics.SOP as SOP
>>> data Bar = A Int Bool | B String Double
>>> instance SOP.Generic Bar
>>> 'view' 'gSOP' (A 10 True)
'InL' (10 ::< True ::< Ø)
>>> 'view' 'gSOP' (B "hello" 3.4)
'InR' ('InL' ("hello" ::< 3.4 ::< Ø))
>>> 'review' 'gTuple' ('InL' (15 ::< False ::< Ø))
A 15 False
>>> 'review' 'gTuple' ('InR' ('InL' ("bye" ::< 9.8 ::< Ø)))
B "bye" 9.8

sum1 :: Iso' (Sum f '[a]) (f a)

An iso between a single-type Sum and the single type.

resum1 :: Iso' (f a) (Sum f '[a])

An iso between a single type and a single-type Sum.