Prelude.Backprop.Explicit

Description

Provides "explicit" versions of all of the functions in Prelude.Backprop. Instead of relying on a Backprop instance, allows you to manually provide zero, add, and one on a per-value basis.

Since: 0.2.0.0

Foldable and Traversable

sum :: (Foldable t, Functor t, Num a, Reifies s W) => AddFunc (t a) -> ZeroFunc a -> BVar s (t a) -> BVar s a

sum, but taking explicit add and zero.

product :: (Foldable t, Functor t, Fractional a, Reifies s W) => AddFunc (t a) -> ZeroFunc a -> BVar s (t a) -> BVar s a

product, but taking explicit add and zero.

length :: (Foldable t, Num b, Reifies s W) => AddFunc (t a) -> ZeroFunc (t a) -> ZeroFunc b -> BVar s (t a) -> BVar s b

length, but taking explicit add and zero.

minimum :: (Foldable t, Functor t, Ord a, Reifies s W) => AddFunc (t a) -> ZeroFunc a -> BVar s (t a) -> BVar s a

minimum, but taking explicit add and zero.

maximum :: (Foldable t, Functor t, Ord a, Reifies s W) => AddFunc (t a) -> ZeroFunc a -> BVar s (t a) -> BVar s a

maximum, but taking explicit add and zero.

traverse :: (Traversable t, Applicative f, Foldable f, Reifies s W) => AddFunc a -> AddFunc b -> AddFunc (t b) -> ZeroFunc a -> ZeroFunc b -> (BVar s a -> f (BVar s b)) -> BVar s (t a) -> BVar s (f (t b))

traverse, but taking explicit add and zero.

See documentation for collectVar for information the API change in v0.2.3 that removed the ZeroFunc (t b) and ZeroFunc (f (t b)) parameters.

toList :: (Traversable t, Reifies s W) => AddFunc a -> ZeroFunc a -> BVar s (t a) -> [BVar s a]

length, but taking explicit add and zero.

Since: 0.2.2.0

mapAccumL :: (Traversable t, Reifies s W) => AddFunc b -> AddFunc c -> ZeroFunc b -> ZeroFunc c -> (BVar s a -> BVar s b -> (BVar s a, BVar s c)) -> BVar s a -> BVar s (t b) -> (BVar s a, BVar s (t c))

mapAccumL, but taking explicit add and zero.

See documentation for collectVar for information the API change in v0.2.3 that removed the ZeroFunc (t c) parameter.

Since: 0.2.2.0

mapAccumR :: (Traversable t, Reifies s W) => AddFunc b -> AddFunc c -> ZeroFunc b -> ZeroFunc c -> (BVar s a -> BVar s b -> (BVar s a, BVar s c)) -> BVar s a -> BVar s (t b) -> (BVar s a, BVar s (t c))

mapAccumR, but taking explicit add and zero.

See documentation for collectVar for information the API change in v0.2.3 that removed the ZeroFunc (t c) parameter.

Since: 0.2.2.0

Functor and Applicative

fmap :: (Traversable f, Reifies s W) => AddFunc a -> AddFunc b -> ZeroFunc a -> ZeroFunc b -> (BVar s a -> BVar s b) -> BVar s (f a) -> BVar s (f b)

fmap, but taking explicit add and zero.

See documentation for collectVar for information the API change in v0.2.3 that removed the ZeroFunc (f b) parameter.

pure :: (Foldable t, Applicative t, Reifies s W) => AddFunc a -> ZeroFunc a -> ZeroFunc (t a) -> BVar s a -> BVar s (t a)

pure, but taking explicit add and zero.

liftA2 :: (Traversable f, Applicative f, Reifies s W) => AddFunc a -> AddFunc b -> AddFunc c -> ZeroFunc a -> ZeroFunc b -> ZeroFunc c -> (BVar s a -> BVar s b -> BVar s c) -> BVar s (f a) -> BVar s (f b) -> BVar s (f c)

liftA2, but taking explicit add and zero.

See documentation for collectVar for information the API change in v0.2.3 that removed the ZeroFunc (f c) parameter.

liftA3 :: (Traversable f, Applicative f, Reifies s W) => AddFunc a -> AddFunc b -> AddFunc c -> AddFunc d -> ZeroFunc a -> ZeroFunc b -> ZeroFunc c -> ZeroFunc d -> (BVar s a -> BVar s b -> BVar s c -> BVar s d) -> BVar s (f a) -> BVar s (f b) -> BVar s (f c) -> BVar s (f d)

liftA3, but taking explicit add and zero.

See documentation for collectVar for information the API change in v0.2.3 that removed the ZeroFunc (f d) parameter.

Misc

fromIntegral :: (Integral a, Integral b, Reifies s W) => AddFunc a -> ZeroFunc b -> BVar s a -> BVar s b

fromIntegral, but taking explicit add and zero.

Since: 0.2.1.0

realToFrac :: (Fractional a, Real a, Fractional b, Real b, Reifies s W) => AddFunc a -> ZeroFunc b -> BVar s a -> BVar s b

realToFrac, but taking explicit add and zero.

Since: 0.2.1.0

coerce :: Coercible a b => BVar s a -> BVar s b

Coerce items inside a BVar.