Prelude.Backprop.Num

Description

Provides the exact same API as Prelude.Backprop, except requiring Num instances for all types involved instead of Backprop instances.

Since: 0.2.0.0

Foldable and Traversable

sum :: forall t a s. (Foldable t, Functor t, Num (t a), Num a, Reifies s W) => BVar s (t a) -> BVar s a

sum, but with Num constraints instead of Backprop constraints.

product :: forall t a s. (Foldable t, Functor t, Num (t a), Fractional a, Reifies s W) => BVar s (t a) -> BVar s a

product, but with Num constraints instead of Backprop constraints.

length :: forall t a b s. (Foldable t, Num (t a), Num b, Reifies s W) => BVar s (t a) -> BVar s b

length, but with Num constraints instead of Backprop constraints.

minimum :: forall t a s. (Foldable t, Functor t, Num a, Ord a, Num (t a), Reifies s W) => BVar s (t a) -> BVar s a

minimum, but with Num constraints instead of Backprop constraints.

maximum :: forall t a s. (Foldable t, Functor t, Num a, Ord a, Num (t a), Reifies s W) => BVar s (t a) -> BVar s a

maximum, but with Num constraints instead of Backprop constraints.

traverse :: forall t f a b s. (Traversable t, Applicative f, Foldable f, Num a, Num b, Num (f (t b)), Num (t b), Reifies s W) => (BVar s a -> f (BVar s b)) -> BVar s (t a) -> BVar s (f (t b))

traverse, but with Num constraints instead of Backprop constraints.

toList :: (Traversable t, Num a, Reifies s W) => BVar s (t a) -> [BVar s a]

toList, but with Num constraints instead of Backprop constraints.

Since: 0.2.2.0

mapAccumL :: (Traversable t, Num b, Num c, Num (t c), Reifies s W) => (BVar s a -> BVar s b -> (BVar s a, BVar s c)) -> BVar s a -> BVar s (t b) -> (BVar s a, BVar s (t c))

mapAccumL, but with Num constraints instead of Backprop constraints.

Since: 0.2.2.0

mapAccumR :: (Traversable t, Num b, Num c, Num (t c), Reifies s W) => (BVar s a -> BVar s b -> (BVar s a, BVar s c)) -> BVar s a -> BVar s (t b) -> (BVar s a, BVar s (t c))

mapAccumR, but with Num constraints instead of Backprop constraints.

Since: 0.2.2.0

Functor and Applicative

fmap :: forall f a b s. (Traversable f, Num a, Num b, Num (f b), Reifies s W) => (BVar s a -> BVar s b) -> BVar s (f a) -> BVar s (f b)

fmap, but with Num constraints instead of Backprop constraints.

(<$>) :: forall f a b s. (Traversable f, Num a, Num b, Num (f b), Reifies s W) => (BVar s a -> BVar s b) -> BVar s (f a) -> BVar s (f b)

Alias for fmap.

pure :: forall t a s. (Foldable t, Applicative t, Num (t a), Num a, Reifies s W) => BVar s a -> BVar s (t a)

pure, but with Num constraints instead of Backprop constraints.

liftA2 :: forall f a b c s. (Traversable f, Applicative f, Num a, Num b, Num c, Num (f c), Reifies s W) => (BVar s a -> BVar s b -> BVar s c) -> BVar s (f a) -> BVar s (f b) -> BVar s (f c)

liftA2, but with Num constraints instead of Backprop constraints.

liftA3 :: forall f a b c d s. (Traversable f, Applicative f, Num a, Num b, Num c, Num d, Num (f d), Reifies s W) => (BVar s a -> BVar s b -> BVar s c -> BVar s d) -> BVar s (f a) -> BVar s (f b) -> BVar s (f c) -> BVar s (f d)

liftA3, but with Num constraints instead of Backprop constraints.

Misc

fromIntegral :: (Integral a, Integral b, Reifies s W) => BVar s a -> BVar s b

fromIntegral, but with Num constraints instead of Backprop constraints.

Since: 0.2.1.0

realToFrac :: (Fractional a, Real a, Fractional b, Real b, Reifies s W) => BVar s a -> BVar s b

realToFrac, but with Num constraints instead of Backprop constraints.

Since: 0.2.1.0

coerce :: forall a b s. Coercible a b => BVar s a -> BVar s b

Coerce items inside a BVar.