Numeric.Backprop

Description

Provides the BP monad and the BVar type; after manipulating BVars (inputs to your function) to produce a result, the library tracks internal data dependencies, which are used to perform back-propagation (reverse-mode automatic differentiation) to calculate the gradient of the output with respect to the inputs.

Similar to automatic differentiation from the ad library and Numeric.AD.Mode.Reverse, except for a few key differences:

Most importantly, this library implements heterogeneous back-propagation, so you can manipulate values of different types (like different matrix and vector types, and product and sum types). This is essential for things like back-propagation for neural networks.

This module allows you to explicitly build your data dependency graph if you wish, which allows the library to perform optimizations and reduce extra allocation, which may or may not provide advantages over Numeric.AD.Mode.Reverse's unsafePerformIO-based implicit graph building.

See the README for more information and links to demonstrations and tutorials. If you want to plunge right in, you can also look directly at the main types, BP, BPOp, BVar, Op, and the main functions, backprop and opVar.

Note that every type involved has to be an instance of Num. This is because gradients all need to be "summable" (which is implemented using sum and +), and we also need to able to generate gradients of '1' and '0'.