Numeric.Backprop

Types

data BP s rs b

A Monad allowing you to explicitly build hetereogeneous data dependency graphs and that the library can perform backpropagation on.

A BP s rs b is a BP action that uses an environment of rs returning a b. When "run", it will compute a gradient that is a tuple of rs. (The phantom parameter s is used to ensure that any BVars aren't leaked out of the monad)

Note that you can only "run" a BP s rs that produces a BVar -- that is, things of the form

BP s rs (BVar rs a)

The above is a BP action that returns a BVar containing an a. When this is run, it'll produce a result of type a and a gradient of that is a tuple of rs. (This form has a type synonym, BPOp, for convenience)

For example, a BP s '[ Int, Double, Double ] is a monad that represents a computation with an Int, Double, and Double as inputs. And, if you ran a

BP s '[ Int, Double, Double ] (BVar '[ Int, Double, Double ] Double)

Or, using the BPOp type synonym:

BPOp s '[ Int, Double, Double ] (BVar rs Double)

with backprop or gradBPOp, it'll return a gradient on the inputs (Int, Double, and Double) and produce a value of type Double.

type BPOp s rs a = BP s rs (BVar s rs a)

A handy type synonym representing a BP action that returns a BVar. This is handy because this is the form of BP actions that backprop and gradBPOp (etc.) expects.

A value of type:

BPOp s rs a

is an action that takes an input environment of rs and produces a BVar containing a value of type a. Because it returns a BVar, the library can track the data dependencies between the BVar and the input environment and perform backpropagation.

See documentation for BP for an explanation of the phantom type parameter s.

type BPOpI s rs a = Prod (BVar s rs) rs -> BVar s rs a

An "implicit" operation on BVars that can be backpropagated. A value of type:

BPOpI s rs a

takes a bunch of BVars containg rs and uses them to (purely) produce a BVar containing an a.

If you are exclusively doing implicit backpropagation by combining BVars and using BPOpIs, you are probably better off just importing Numeric.Backprop.Implicit, which provides better tools. This type synonym exists in Numeric.Backprop just for the implicitly function, which can convert "implicit" backprop functions like a BPOpI s rs a into an "explicit" graph backprop function, a BPOp s rs a.

data BVar :: Type -> [Type] -> Type -> Type

The basic unit of manipulation inside BP. Instead of directly working with values, you work with BVars contating those values. When you work with a BVar, the backprop library can keep track of what values refer to which other values, and so can perform backpropagation to compute gradients.

A BVar s rs a refers to a value of type a, with an environment of values of the types rs. The phantom parameter s is used to ensure that stray BVars don't leak outside of the backprop process.

(That is, if you're using implicit backprop, it ensures that you interact with BVars in a polymorphic way. And, if you're using implicit backprop, it ensures that a BVar s rs a never leaves the BP s rs that it was created in.)

BVars have Num, Fractional, Floating, etc. instances, so they can be manipulated using polymorphic functions and numeric functions in Haskell. You can add them, subtract them, etc., in "implicit" backprop style.

(However, note that if you directly manipulate BVars using those instances or using liftR, it delays evaluation, so every usage site has to re-compute the result/create a new node. If you want to re-use a BVar you created using '(+)' or '(-)' or liftR, use bindRef to force it first.)

type Op as a = forall m. Monad m => OpM m as a

type OpB s as a = OpM (ST s) as a

A type synonym for OpM specialized to what the backprop library uses to perform backpropagation.

An

OpB s rs a

represents a differentiable function that takes a tuple of rs and produces an a a, which can be run on BVar s and also inside BP ss. For example, an OpB s '[ Int, Double ] Bool takes an Int and a Double and produces a Bool, and does it in a differentiable way.

Note that OpB is a superset of Op, so, if you see any function that expects an OpB (like opRef' and ~$, for example), you can give them an Op, as well.

You can think of OpB as a superclass/parent class of Op in this sense, and of Op as a subclass of OpB.

BP

Backprop

backprop :: forall rs a. Every Num rs => (forall s. BPOp s rs a) -> Tuple rs -> (a, Tuple rs)

evalBPOp :: Every Num rs => (forall s. BPOp s rs a) -> Tuple rs -> a

gradBPOp :: Every Num rs => (forall s. BPOp s rs a) -> Tuple rs -> Tuple rs

bpOp :: (Every Num as, Known Length as) => BPOp s as a -> OpB s as a

backprop' :: Prod Summer rs -> Prod Unity rs -> (forall s. BPOp s rs a) -> Tuple rs -> (a, Tuple rs)

evalBPOp' :: Prod Summer rs -> Prod Unity rs -> (forall s. BPOp s rs a) -> Tuple rs -> a

gradBPOp' :: Prod Summer rs -> Prod Unity rs -> (forall s. BPOp s rs a) -> Tuple rs -> Tuple rs

bpOp' :: Prod Summer as -> Prod Unity as -> BPOp s as a -> OpB s as a

Inputs

withInps :: Known Length rs => (Prod (BVar s rs) rs -> BP s rs a) -> BP s rs a

implicitly :: Known Length rs => BPOpI s rs a -> BPOp s rs a

withInps' :: Length rs -> (Prod (BVar s rs) rs -> BP s rs a) -> BP s rs a

implicitly' :: Length rs -> BPOpI s rs a -> BPOp s rs a

Refs

constRef :: a -> BVar s rs a

inpRef :: Index rs a -> BVar s rs a

inpRefs :: Known Length rs => Prod (BVar s rs) rs

bindRef :: Num a => BVar s rs a -> BP s rs (BVar s rs a)

inpRefs' :: Length rs -> Prod (BVar s rs) rs

bindRef' :: Summer a -> BVar s rs a -> BP s rs (BVar s rs a)

From Ops

opRef :: Num a => Prod (BVar s rs) as -> OpB s as a -> BP s rs (BVar s rs a)

(~$) :: Num a => OpB s as a -> Prod (BVar s rs) as -> BP s rs (BVar s rs a) infixr 1

opRef1 :: Num b => BVar s rs a -> OpB s '[a] b -> BP s rs (BVar s rs b)

opRef2 :: Num c => BVar s rs a -> BVar s rs b -> OpB s '[a, b] c -> BP s rs (BVar s rs c)

opRef3 :: Num d => BVar s rs a -> BVar s rs b -> BVar s rs c -> OpB s '[a, b, c] d -> BP s rs (BVar s rs d)

(-$) :: (Every Num as, Known Length as, Num a) => BPOp s as a -> Prod (BVar s rs) as -> BPOp s rs a infixr 1

opRef' :: forall s rs as a. Summer a -> Prod (BVar s rs) as -> OpB s as a -> BP s rs (BVar s rs a)

opRef1' :: Summer b -> BVar s rs a -> OpB s '[a] b -> BP s rs (BVar s rs b)

opRef2' :: Summer c -> BVar s rs a -> BVar s rs b -> OpB s '[a, b] c -> BP s rs (BVar s rs c)

opRef3' :: Summer d -> BVar s rs a -> BVar s rs b -> BVar s rs c -> OpB s '[a, b, c] d -> BP s rs (BVar s rs d)

Ref manipulation

As parts

partsRef :: forall s rs bs b. (Every Num bs, Known Length bs) => Iso' b (Tuple bs) -> BVar s rs b -> BP s rs (Prod (BVar s rs) bs)

(#<~) :: (Every Num bs, Known Length bs) => Iso' b (Tuple bs) -> BVar s rs b -> BP s rs (Prod (BVar s rs) bs) infixr 1

withParts :: (Every Num bs, Known Length bs) => Iso' b (Tuple bs) -> BVar s rs b -> (Prod (BVar s rs) bs -> BP s rs a) -> BP s rs a

splitRefs :: forall s rs as. (Every Num as, Known Length as) => BVar s rs (Tuple as) -> BP s rs (Prod (BVar s rs) as)

gSplit :: (Every Num bs, Known Length bs, Generic b, Code b ~ '[bs]) => BVar s rs b -> BP s rs (Prod (BVar s rs) bs)

partsRef' :: forall s rs bs b. Prod Summer bs -> Prod Unity bs -> Iso' b (Tuple bs) -> BVar s rs b -> BP s rs (Prod (BVar s rs) bs)

withParts' :: Prod Summer bs -> Prod Unity bs -> Iso' b (Tuple bs) -> BVar s rs b -> (Prod (BVar s rs) bs -> BP s rs a) -> BP s rs a

splitRefs' :: forall s rs as. Prod Summer as -> Prod Unity as -> BVar s rs (Tuple as) -> BP s rs (Prod (BVar s rs) as)

gSplit' :: (Generic b, Code b ~ '[bs]) => Prod Summer bs -> Prod Unity bs -> BVar s rs b -> BP s rs (Prod (BVar s rs) bs)

As sums

choicesRef :: forall s rs bs b. (Every Num bs, Known Length bs) => Iso' b (Sum I bs) -> BVar s rs b -> BP s rs (Sum (BVar s rs) bs)

choicesRef' :: forall s rs bs b. Prod Summer bs -> Prod Unity bs -> Iso' b (Sum I bs) -> BVar s rs b -> BP s rs (Sum (BVar s rs) bs)

As sums of products

sopRef :: forall s rs bss b. (Known Length bss, Every (Every Num ∧ Known Length) bss) => Iso' b (Sum Tuple bss) -> BVar s rs b -> BP s rs (Sum (Prod (BVar s rs)) bss)

gSplits :: forall s rs b. (Generic b, Known Length (Code b), Every (Every Num ∧ Known Length) (Code b)) => BVar s rs b -> BP s rs (Sum (Prod (BVar s rs)) (Code b))

sopRef' :: forall s rs bss b. Prod (Prod Summer) bss -> Prod (Prod Unity) bss -> Iso' b (Sum Tuple bss) -> BVar s rs b -> BP s rs (Sum (Prod (BVar s rs)) bss)

gSplits' :: forall s rs b. Generic b => Prod (Prod Summer) (Code b) -> Prod (Prod Unity) (Code b) -> BVar s rs b -> BP s rs (Sum (Prod (BVar s rs)) (Code b))

Transforming BP

internally :: forall s rs bs b a. (Every Num bs, Known Length bs, Num a) => Iso' b (Tuple bs) -> BVar s rs b -> BP s bs (BVar s bs a) -> BP s rs (BVar s rs a)

generically :: forall s rs bs b a. (Num a, Every Num bs, Known Length bs, Generic b, Code b ~ '[bs]) => BVar s rs b -> BP s bs (BVar s bs a) -> BP s rs (BVar s rs a)

internally' :: forall s rs bs b a. Prod Summer bs -> Prod Unity bs -> Summer a -> Iso' b (Tuple bs) -> BVar s rs b -> BP s bs (BVar s bs a) -> BP s rs (BVar s rs a)

generically' :: forall s rs bs b a. (Generic b, Code b ~ '[bs]) => Prod Summer bs -> Prod Unity bs -> Summer a -> BVar s rs b -> BP s bs (BVar s bs a) -> BP s rs (BVar s rs a)

Combining

liftR :: OpB s as a -> Prod (BVar s rs) as -> BVar s rs a

liftR1 :: OpB s '[a] b -> BVar s rs a -> BVar s rs b

liftR2 :: OpB s '[a, b] c -> BVar s rs a -> BVar s rs b -> BVar s rs c

liftR3 :: OpB s '[a, b, c] d -> BVar s rs a -> BVar s rs b -> BVar s rs c -> BVar s rs d

Op

op1 :: Num a => (forall s. AD s (Forward a) -> AD s (Forward a)) -> Op '[a] a

op2 :: Num a => (forall s. Reifies s Tape => Reverse s a -> Reverse s a -> Reverse s a) -> Op '[a, a] a

op3 :: Num a => (forall s. Reifies s Tape => Reverse s a -> Reverse s a -> Reverse s a -> Reverse s a) -> Op '[a, a, a] a

opN :: (Num a, Known Nat n) => (forall s. Reifies s Tape => Vec n (Reverse s a) -> Reverse s a) -> Op (Replicate n a) a

op1' :: (a -> (b, Maybe b -> a)) -> Op '[a] b

op2' :: (a -> b -> (c, Maybe c -> (a, b))) -> Op '[a, b] c

op3' :: (a -> b -> c -> (d, Maybe d -> (a, b, c))) -> Op '[a, b, c] d

opN' :: Known Nat n => (Vec n a -> (b, Maybe b -> Vec n a)) -> Op (Replicate n a) b

Utility

data Prod k f a :: forall k. (k -> *) -> [k] -> * where

Constructors

Ø :: Prod k f ([] k)

(:<) :: Prod k f ((:) k a1 as) infixr 5

pattern (:>) :: forall k f a b. f a -> f b -> Prod k f ((:) k a ((:) k b ([] k))) infix 6

Construct a two element Prod. Since the precedence of (:>) is higher than (:<), we can conveniently write lists like:

>>> a :< b :> c

Which is identical to:

>>> a :< b :< c :< Ø

only :: f a -> Prod k f ((:) k a ([] k))

Build a singleton Prod.

head' :: Prod k f ((:<) k a as) -> f a

type Tuple = Prod * I

A Prod of simple Haskell types.

pattern (::<) :: forall a as. a -> Tuple as -> Tuple ((:<) * a as) infixr 5

Cons onto a Tuple.

only_ :: a -> Tuple ((:) * a ([] *))

Singleton Tuple.

newtype Summer a

Instructions on how to "sum" a list of values of a given type. Basically used as an explicit witness for a Num instance.

For most types, the only meaningful value of type Summer a is Summer sum. However, using Summer lets us use BP with types that are not instances of Num. Any type can be used, as long as you provide a way to "sum" it!

For most of the functions in this library, you can completely ignore this, as they will be generated automatically. You only need to work with this directly if you want to use custom types that aren't instances of Num with this library.

If 'Num a' is satisfied, one can create the canonical Summer using known :: Num a => Summer a.

Constructors

Summer

Fields

runSummer :: [a] -> a

summers :: (Every Num as, Known Length as) => Prod Summer as

If all the types in as are instances of Num, generate a Prod Summer as, or a tuple of Summers for every type in as.

summers' :: Every Num as => Length as -> Prod Summer as

Like summers, but requiring an explicit witness for the number of types in the list as.

newtype Unity a

A canonical "unity" (the multiplicative identity) for a given type. Basically used as an explicit witness for a Num instance.

For most types, the only meaningful value of type Unity a is Unity 1'. However, using Unity lets us use BP with types that are not instances of Num. Any type can be used, as long as you provide a way to get a multiplicative identity in it!

For most of the functions in this library, you can completely ignore this, as they will be generated automatically. You only need to work with this directly if you want to use custom types that aren't instances of Num with this library.

If 'Num a' is satisfied, one can create the canonical Unity using known :: Num a => Unity a.

Constructors

Unity

Fields

getUnity :: a

unities :: (Every Num as, Known Length as) => Prod Unity as

If all the types in as are instances of Num, generate a Prod Unity as, or a tuple of Unitys for every type in as.

unities' :: Every Num as => Length as -> Prod Unity as

Like unities, but requiring an explicit witness for the number of types in the list as.