Numeric.Backprop.Class

Description

Provides the Backprop typeclass, a class for values that can be used for backpropagation.

This class replaces the old (version 0.1) API relying on Num.

Since: backprop-0.2.0.0

Backpropagatable types

class Backprop a where

Class of values that can be backpropagated in general.

For instances of Num, these methods can be given by zeroNum, addNum, and oneNum. There are also generic options given in Numeric.Backprop.Class for functors, IsList instances, and Generic instances.

instance Backprop Double where
    zero = zeroNum
    add = addNum
    one = oneNum

If you leave the body of an instance declaration blank, GHC Generics will be used to derive instances if the type has a single constructor and each field is an instance of Backprop.

To ensure that backpropagation works in a sound way, should obey the laws:

identity

add x (zero y) = x

add (zero x) y = y

Also implies preservation of information, making zipWith (+) an illegal implementation for lists and vectors.

This is only expected to be true up to potential "extra zeroes" in x and y in the result.

commutativity

add x y = add y x

associativity

add x (add y z) = add (add x y) z

idempotence

zero . zero = zero

one . one = one

unital

one = gradBP id

Note that not all values in the backpropagation process needs all of these methods: Only the "final result" needs one, for example. These are all grouped under one typeclass for convenience in defining instances, and also to talk about sensible laws. For fine-grained control, use the "explicit" versions of library functions (for example, in Numeric.Backprop.Explicit) instead of Backprop based ones.

This typeclass replaces the reliance on Num of the previous API (v0.1). Num is strictly more powerful than Backprop, and is a stronger constraint on types than is necessary for proper backpropagating. In particular, fromInteger is a problem for many types, preventing useful backpropagation for lists, variable-length vectors (like Data.Vector) and variable-size matrices from linear algebra libraries like hmatrix and accelerate.

Since: backprop-0.2.0.0

Methods

zero :: a -> a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

Should be idempotent:

zero . zero = zero

Should be as lazy as possible. This behavior is observed for all instances provided by this library.

See zeroNum for a pre-built definition for instances of Num and zeroFunctor for a definition for instances of Functor. If left blank, will automatically be genericZero, a pre-built definition for instances of Generic whose fields are all themselves instances of Backprop.

add :: a -> a -> a

Add together two values of a type. To combine contributions of gradients, so should be information-preserving:

add x (zero y) = x

add (zero x) y = y

Should be as strict as possible. This behavior is observed for all instances provided by this library.

See addNum for a pre-built definition for instances of Num and addIsList for a definition for instances of IsList. If left blank, will automatically be genericAdd, a pre-built definition for instances of Generic with one constructor whose fields are all themselves instances of Backprop.

one :: a -> a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

As the library uses it, the most important law is:

one = gradBP id

That is, one x is the gradient of the identity function with respect to its input.

Ideally should be idempotent:

one . one = one

Should be as lazy as possible. This behavior is observed for all instances provided by this library.

See oneNum for a pre-built definition for instances of Num and oneFunctor for a definition for instances of Functor. If left blank, will automatically be genericOne, a pre-built definition for instances of Generic whose fields are all themselves instances of Backprop.

zero :: (Generic a, GZero (Rep a)) => a -> a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

Should be idempotent:

zero . zero = zero

Should be as lazy as possible. This behavior is observed for all instances provided by this library.

See zeroNum for a pre-built definition for instances of Num and zeroFunctor for a definition for instances of Functor. If left blank, will automatically be genericZero, a pre-built definition for instances of Generic whose fields are all themselves instances of Backprop.

add :: (Generic a, GAdd (Rep a)) => a -> a -> a

Add together two values of a type. To combine contributions of gradients, so should be information-preserving:

add x (zero y) = x

add (zero x) y = y

Should be as strict as possible. This behavior is observed for all instances provided by this library.

See addNum for a pre-built definition for instances of Num and addIsList for a definition for instances of IsList. If left blank, will automatically be genericAdd, a pre-built definition for instances of Generic with one constructor whose fields are all themselves instances of Backprop.

one :: (Generic a, GOne (Rep a)) => a -> a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

As the library uses it, the most important law is:

one = gradBP id

That is, one x is the gradient of the identity function with respect to its input.

Ideally should be idempotent:

one . one = one

Should be as lazy as possible. This behavior is observed for all instances provided by this library.

See oneNum for a pre-built definition for instances of Num and oneFunctor for a definition for instances of Functor. If left blank, will automatically be genericOne, a pre-built definition for instances of Generic whose fields are all themselves instances of Backprop.

Derived methods

zeroNum :: Num a => a -> a

zero for instances of Num.

Is lazy in its argument.

addNum :: Num a => a -> a -> a

add for instances of Num.

oneNum :: Num a => a -> a

one for instances of Num.

Is lazy in its argument.

zeroVec :: (Vector v a, Backprop a) => v a -> v a

zero for instances of Vector.

addVec :: (Vector v a, Backprop a) => v a -> v a -> v a

add for instances of Vector. Automatically pads the end of the shorter vector with zeroes.

oneVec :: (Vector v a, Backprop a) => v a -> v a

one for instances of Vector.

zeroVecNum :: (Vector v a, Num a) => v a -> v a

zero for instances of Vector when the contained type is an instance of Num. Is potentially more performant than zeroVec when the vectors are larger.

See NumVec for a Backprop instance for Vector instances that uses this for zero.

Since: backprop-0.2.4.0

oneVecNum :: (Vector v a, Num a) => v a -> v a

one for instances of Vector when the contained type is an instance of Num. Is potentially more performant than oneVec when the vectors are larger.

See NumVec for a Backprop instance for Vector instances that uses this for one.

Since: backprop-0.2.4.0

zeroFunctor :: (Functor f, Backprop a) => f a -> f a

zero for Functor instances.

addIsList :: (IsList a, Backprop (Item a)) => a -> a -> a

add for instances of IsList. Automatically pads the end of the "shorter" value with zeroes.

addAsList

Arguments

:: Backprop b

=> (a -> [b])
convert to list (should form isomorphism)

-> ([b] -> a)
convert from list (should form isomorphism)

-> a

-> a

-> a

add for types that are isomorphic to a list. Automatically pads the end of the "shorter" value with zeroes.

oneFunctor :: (Functor f, Backprop a) => f a -> f a

one for instances of Functor.

genericZero :: (Generic a, GZero (Rep a)) => a -> a

zero using GHC Generics; works if all fields are instances of Backprop.

genericAdd :: (Generic a, GAdd (Rep a)) => a -> a -> a

add using GHC Generics; works if all fields are instances of Backprop, but only for values with single constructors.

genericOne :: (Generic a, GOne (Rep a)) => a -> a

one using GHC Generics; works if all fields are instaces of Backprop.

Newtype

newtype ABP f a

A newtype wrapper over an f a for Applicative f that gives a free Backprop instance (as well as Num etc. instances).

Useful for performing backpropagation over functions that require some monadic context (like IO) to perform.

Since: backprop-0.2.1.0

Constructors

ABP

Fields

runABP :: f a

newtype NumBP a

A newtype wrapper over an instance of Num that gives a free Backprop instance.

Useful for things like DerivingVia, or for avoiding orphan instances.

Since: backprop-0.2.1.0

Constructors

NumBP

Fields

runNumBP :: a

newtype NumVec v a

Newtype wrapper around a v a for Vector v a, that gives a more efficient Backprop instance for long vectors when a is an instance of Num. The normal Backprop instance for vectors will map zero or one over all items; this instance will completely ignore the contents of the original vector and instead produce a new vector of the same length, with all 0 or 1 using the Num instance of a (essentially using zeroVecNum and oneVecNum instead of zeroVec and oneVec).

add is essentially the same as normal, but using + instead of the type's add.

Since: backprop-0.2.4.0

Constructors

NumVec

Fields

runNumVec :: v a

Generics

class GZero f

Helper class for automatically deriving zero using GHC Generics.

Minimal complete definition

gzero

class GAdd f

Helper class for automatically deriving add using GHC Generics.

Minimal complete definition

gadd

class GOne f

Helper class for automatically deriving one using GHC Generics.

Minimal complete definition

gone