Numeric.Backprop.Class

Backpropagatable types

class Backprop a where

Class of values that can be backpropagated in general.

For instances of Num, these methods can be given by zeroNum, addNum, and oneNum. There are also generic options given in Numeric.Backprop.Class for functors, IsList instances, and Generic instances.

If you leave the body of an instance declaration blank, GHC Generics will be used to derive instances if the type has a single constructor and each field is an instance of Backprop.

To ensure that backpropagation works in a sound way, should obey the laws:

identity

add x (zero y) = x

add (zero x) y = y

commutativity

add x y = add y x

associativity

add x (add y z) = add (add x y) z

idempotence

zero . zero = zero

one . one = one

Note that not all values in the backpropagation process needs all of these methods: Only the "final result" needs one, for example. These are all grouped under one typeclass for convenience in defining instances, and also to talk about sensible laws. For fine-grained control, use the "explicit" versions of library functions (for example, in Numeric.Backprop.Explicit) instead of Backprop based ones.

Methods

zero :: a -> a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

add :: a -> a -> a

Add together two values of a type. To combine contributions of gradients, so should ideally be information-preserving.

one :: a -> a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

zero :: (Generic a, GZero (Rep a)) => a -> a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

add :: (Generic a, GAdd (Rep a)) => a -> a -> a

Add together two values of a type. To combine contributions of gradients, so should ideally be information-preserving.

one :: (Generic a, GOne (Rep a)) => a -> a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

Derived methods

zeroNum :: Num a => a -> a

addNum :: Num a => a -> a -> a

oneNum :: Num a => a -> a

zeroVec :: (Vector v a, Backprop a) => v a -> v a

addVec :: (Vector v a, Backprop a) => v a -> v a -> v a

oneVec :: (Vector v a, Backprop a) => v a -> v a

zeroFunctor :: (Functor f, Backprop a) => f a -> f a

addIsList :: (IsList a, Backprop (Item a)) => a -> a -> a

addAsList :: Backprop b => (a -> [b]) -> ([b] -> a) -> a -> a -> a

oneFunctor :: (Functor f, Backprop a) => f a -> f a

genericZero :: (Generic a, GZero (Rep a)) => a -> a

genericAdd :: (Generic a, GAdd (Rep a)) => a -> a -> a

genericOne :: (Generic a, GOne (Rep a)) => a -> a

Usable by backprop

newtype ZeroFunc a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

Should be idempotent: Applying the function twice is the same as applying it just once.

Each type should ideally only have one ZeroFunc. This coherence constraint is given by the typeclass Zero.

Constructors

ZF

Fields

runZF :: a -> a

zeroFunc :: Backprop a => ZeroFunc a

zeroFuncs :: (Every Backprop as, Known Length as) => Prod ZeroFunc as

zfNum :: Num a => ZeroFunc a

If a type has a Num instance, this is the canonical ZeroFunc.

zfNums :: (Every Num as, Known Length as) => Prod ZeroFunc as

ZeroFuncs for every item in a type level list based on their Num instances

newtype AddFunc a

Add together two values of a type. To combine contributions of gradients, so should ideally be information-preserving.

See laws for Backprop for the laws this should be expected to preserve. Namely, it should be commutative and associative, with an identity for a valid ZeroFunc.

Each type should ideally only have one AddFunc. This coherence constraint is given by the typeclass Add.

Constructors

AF

Fields

runAF :: a -> a -> a

addFunc :: Backprop a => AddFunc a

addFuncs :: (Every Backprop as, Known Length as) => Prod AddFunc as

afNum :: Num a => AddFunc a

If a type has a Num instance, this is the canonical AddFunc.

afNums :: (Every Num as, Known Length as) => Prod AddFunc as

ZeroFuncs for every item in a type level list based on their Num instances

newtype OneFunc a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

Should be idempotent: Applying the function twice is the same as applying it just once.

Each type should ideally only have one ZeroFunc. This coherence constraint is given by the typeclass One.

Constructors

OF

Fields

runOF :: a -> a

oneFunc :: Backprop a => OneFunc a

oneFuncs :: (Every Backprop as, Known Length as) => Prod OneFunc as

ofNum :: Num a => OneFunc a

If a type has a Num instance, this is the canonical OneFunc.

ofNums :: (Every Num as, Known Length as) => Prod OneFunc as

ZeroFuncs for every item in a type level list based on their Num instances

Generics

class GZero f where

Minimal complete definition

gzero

Methods

gzero :: f t -> f t

class GAdd f where

Minimal complete definition

gadd

Methods

gadd :: f t -> f t -> f t

class GOne f where

Minimal complete definition

gone

Methods

gone :: f t -> f t