Numeric.Backprop.Class

Documentation

class Zero a where

Methods

zero :: a -> a

zero :: Num a => a -> a

newtype ZeroFunc a

"Zero out" all components of a value. For scalar values, this should just be const 0. For vectors and matrices, this should set all components to zero, the additive identity.

Each type should ideally only have one ZeroFunc. This coherence constraint is given by the typeclass Zero.

Constructors

ZF

Fields

runZF :: a -> a

zeroFunc :: Zero a => ZeroFunc a

zeroFuncs :: (Every Zero as, Known Length as) => Prod ZeroFunc as

class Add a where

Methods

add :: a -> a -> a

add :: Num a => a -> a -> a

newtype AddFunc a

Add together two values of a type. To combine contributions of gradients, so should ideally be information-preserving. For any other valid ZeroFunc, should ideally obey:

af zf x y -> runAF af x (runZF zf y) == x
           && runAF af (runZF zf x) y == y

Each type should ideally only have one AddFunc. This coherence constraint is given by the typeclass Add.

Constructors

AF

Fields

runAF :: a -> a -> a

addFunc :: Add a => AddFunc a

addFuncs :: (Every Add as, Known Length as) => Prod AddFunc as

class One a where

Methods

one :: a -> a

one :: Num a => a -> a

newtype OneFunc a

One all components of a value. For scalar values, this should just be const 1. For vectors and matrices, this should set all components to one, the multiplicative identity.

Each type should ideally only have one ZeroFunc. This coherence constraint is given by the typeclass One.

Constructors

OF

Fields

runOF :: a -> a

oneFunc :: One a => OneFunc a

oneFuncs :: (Every One as, Known Length as) => Prod OneFunc as

zeroVec :: (Vector v a, Zero a) => v a -> v a

addVec :: (Vector v a, Add a) => v a -> v a -> v a

oneVec :: (Vector v a, One a) => v a -> v a

zeroFunctor :: (Functor f, Zero a) => f a -> f a

addIsList :: (IsList a, Add (Item a)) => a -> a -> a

oneFunctor :: (Functor f, One a) => f a -> f a