Numeric.Backprop.Op

Description

Provides the Op (and OpM) type and combinators, which represent differentiable functions/operations on values, and are used by the library to perform backpropagation.

Note that Op is a subset or subtype of OpM, and so, any function that expects an OpM m as a (or an OpB s as a) can be given an Op as a and it'll work just fine.

Op

Ops contain information on a function as well as its gradient, but provides that information in a way that allows them to be "chained".

For example, for a function

\[ f : \mathbb{R}^n \rightarrow \mathbb{R} \]

We might want to apply it:

\[ \eqalign{ y &= f(\mathbf{x})\cr z &= g(y) } \]

Now, we might want the gradient \(\nabla z\) with respect to \(\mathbf{x}\), or \(\nabla_\mathbf{x} z\). Explicitly, this is:

\[ \nabla_\mathbf{x} z = \left< \frac{\partial z}{\partial x_1}, \frac{\partial z}{\partial x_2}, \ldots \right> \]

We can compute that by multipling the total derivative of \(z\) with respect to \(y\) (that is, \(\frac{dz}{dy}\) with the gradient of \(f\)):

\[ \eqalign{ \nabla_\mathbf{x} z &= \frac{dz}{dy} \left< \frac{\partial y}{\partial x_1}, \frac{\partial y}{\partial x_2}, \ldots \right>\cr \nabla_\mathbf{x} z &= \frac{dz}{dy} \nabla_\mathbf{x} y } \]

So, to create an Op as a with the Op constructor (or an OpM with the OpM constructor), you give a function that returns a tuple, containing:

An a: The result of the function

An Maybe a -> Tuple as: A function that, when given \(\frac{dz}{dy}\) (in a Just), returns the total gradient \(\nabla_z \mathbf{x}\). If the function is given is given Nothing, then \(\frac{dz}{dy}\) should be taken to be 1. In other words, you would simply need to return \(\nabla_y \mathbf{x}\), unchanged. The "final result" is just \(f(\mathbf{x})\), and not \(g(f(\mathbf{x}))\).

This is done so that Ops can easily be "chained" together, one after the other. If you have an Op for \(f\) and an Op for \(g\), you can compute the gradient of \(f\) knowing that the result target is \(g \circ f\).

Note that end users should probably never be required to construct an Op or OpM explicitly this way. Instead, libraries should provide carefuly pre-constructed ones, or provide ways to generate them automatically (like op1, op2, and op3 here).

type Op as a = forall m. Monad m => OpM m as a

An Op as a is a type synonym over OpM that describes a differentiable function from as to a.

For example, an

Op '[Int, Bool] Double

is a function from an Int and a Bool, returning a Double. It can be differentiated to give a gradient of an Int and a Bool if given a total derivative for a.

See runOp, gradOp, and gradOpWith for examples on how to run it, and Op for instructions on creating it.

This type is abstracted over using the pattern synonym with constructor Op, so you can create one from scratch with it. However, it's simplest to create it using op2', op1', op2', and op3' helper smart constructors And, if your function is a numeric function, they can even be created automatically using op1, op2, op3, and opN with a little help from Numeric.AD from the ad library.

Note that this type is a subset or subtype of OpM. So, if a function ever expects an OpM m as a, you can always provide an Op as a instead.

Many functions in this library will expect an OpM m as a (or an OpB s as a), and in all of these cases, you can provide an Op as a.

pattern Op :: forall as a. (Tuple as -> (a, Maybe a -> Tuple as)) -> Op as a

Construct an Op by giving a function creating the result, and also a continuation on how to create the gradient, given the total derivative of a.

See the module documentation for Numeric.Backprop.Op for more details on the function that this constructor and OpM expect.

newtype OpM m as a

An OpM m as a represents a differentiable (monadic) function from as to a, in the context of a Monad m.

For example, an

OpM IO '[Int, Bool] Double

would be a function that takes an Int and a Bool and returns a Double (in IO). It can be differentiated to give a gradient of an Int and a Bool (in IO) if given the total derivative for a.

Note that an OpM is a superclass of Op, so any function that expects an OpM m as a can also accept an Op as a.

See runOpM, gradOpM, and gradOpWithM for examples on how to run it.

Constructors

OpM (Tuple as -> m (a, Maybe a -> m (Tuple as)))
Construct an OpM by giving a function creating the result, and also a continuation on how to create the gradient, given the total derivative of a.

See the module documentation for Numeric.Backprop.Op for more details on the function that this constructor and Op expect.

data Prod k f a :: forall k. (k -> *) -> [k] -> * where

Constructors

Ø :: Prod k f ([] k)

(:<) :: Prod k f ((:) k a1 as) infixr 5

pattern (:>) :: forall k f a b. f a -> f b -> Prod k f ((:) k a ((:) k b ([] k))) infix 6

Construct a two element Prod. Since the precedence of (:>) is higher than (:<), we can conveniently write lists like:

>>> a :< b :> c

Which is identical to:

>>> a :< b :< c :< Ø

only :: f a -> Prod k f ((:) k a ([] k))

Build a singleton Prod.

head' :: Prod k f ((:<) k a as) -> f a

type Tuple = Prod * I

A Prod of simple Haskell types.

pattern (::<) :: forall a as. a -> Tuple as -> Tuple ((:<) * a as) infixr 5

Cons onto a Tuple.

only_ :: a -> Tuple ((:) * a ([] *))

Singleton Tuple.

Running

Pure

runOp :: Op as a -> Tuple as -> a

Run the function that an Op encodes, to get the result.

>>> runOp (op2 (*)) (3 ::< 5 ::< Ø)
15

gradOp :: Op as a -> Tuple as -> Tuple as

Run the function that an Op encodes, and get its gradient with respect to the result.

>>> gradOp (op2 (*)) (3 ::< 5 ::< Ø)
5 ::< 3 ::< Ø
-- the gradient of x*y is (y, x)

gradOp' :: Op as a -> Tuple as -> (a, Tuple as)

Run the function that an Op encodes, to get the result and also the gradient with respect to the result.

>>> gradOpM' (op2 (*)) (3 ::< 5 ::< Ø) :: IO (Int, Tuple '[Int, Int])
(15, 5 ::< 3 ::< Ø)

gradOpWith

Arguments

:: Op as a
Op to run

-> Tuple as
Inputs to run it with

-> a
The total derivative of the result

-> Tuple as
The gradient

Run the function that an Op encodes, and get its gradient with respect to the total derivative of the result.

See gradOp and the module documentaiton for Numeric.Backprop.Op for more information.

gradOpWith'

Arguments

:: Op as a
Op to run

-> Tuple as
Inputs to run it with

-> Maybe a
If Just, taken as the total derivative of the result. If Nothing, assumes that the result is the final result.

-> Tuple as
The gradient

A combination of gradOp and gradOpWith. The third argument is (optionally) the total derivative the result. Give Nothing and it is assumed that the result is the final result (and the total derivative is 1), and this behaves the same as gradOp. Give Just d and it uses the d as the total derivative of the result, and this behaves like gradOpWith.

See gradOp and the module documentaiton for Numeric.Backprop.Op for more information.

runOp'

Arguments

:: Op as a
Op to run

-> Tuple as
Inputs

-> (a, Maybe a -> Tuple as)
Result, and continuation to get the gradient

A combination of runOp and gradOpWith'. Given an Op and inputs, returns the result of the Op and a continuation that gives its gradient.

The continuation takes the total derivative of the result as input. See documenation for gradOpWith' and module documentation for Numeric.Backprop.Op for more information.

Monadic

runOpM :: Functor m => OpM m as a -> Tuple as -> m a

The monadic version of runOp, for OpMs.

>>> runOpM (op2 (*)) (3 ::< 5 ::< Ø) :: IO Int
15

gradOpM :: Monad m => OpM m as a -> Tuple as -> m (Tuple as)

The monadic version of gradOp, for OpMs.

gradOpM' :: Monad m => OpM m as a -> Tuple as -> m (a, Tuple as)

The monadic version of gradOp', for OpMs.

gradOpWithM

Arguments

:: Monad m

=> OpM m as a
OpM to run

-> Tuple as
Inputs to run it with

-> a
The total derivative of the result

-> m (Tuple as)
the gradient

The monadic version of gradOpWith, for OpMs.

gradOpWithM'

Arguments

:: Monad m

=> OpM m as a
OpM to run

-> Tuple as
Inputs to run it with

-> Maybe a
If Just, taken as the total derivative of the result. If Nothing, assumes that the result is the final result.

-> m (Tuple as)
The gradient

The monadic version of gradOpWith', for OpMs.

runOpM'

Arguments

:: OpM m as a
OpM to run

-> Tuple as
Inputs

-> m (a, Maybe a -> m (Tuple as))
Result, and continuation to get the gradient

A combination of runOpM and gradOpWithM'. Given an OpM and inputs, returns the result of the OpM and a continuation that gives its gradient.

The continuation takes the total derivative of the result as input. See documenation for gradOpWithM' and module documentation for Numeric.Backprop.Op for more information.

Manipulation

composeOp

Arguments

:: (Monad m, Known Length as, Every Num as)

=> Prod (OpM m as) bs
Prod of OpMs taking as and returning different b in bs

-> OpM m bs c
OpM taking eac of the bs from the input Prod.

-> OpM m as c
Composed OpM

Compose OpMs together, similar to .. But, because all OpMs are \(\mathbb{R}^N \rightarrow \mathbb{R}\), this is more like sequence for functions, or liftAN.

That is, given an OpM m as b1, an OpM m as b2, and an OpM m as b3, it can compose them with an OpM m '[b1,b2,b3] c to create an OpM as c.

composeOp'

Arguments

:: Monad m

=> Prod Summer as
Explicit Summers

-> Prod (OpM m as) bs
Prod of OpMs taking as and returning different b in bs

-> OpM m bs c
OpM taking eac of the bs from the input Prod.

-> OpM m as c
Composed OpM

composeOp, but taking explicit Summers, for the situation where the as are not instance of Num.

Creation

op0 :: a -> Op '[] a

Create an Op that takes no inputs and always returns the given value.

There is no gradient, of course (using gradOp will give you an empty tuple), because there is no input to have a gradient of.

>>> gradOp' (op0 10) Ø
(10, Ø)

For a constant Op that takes input and ignores it, see opConst and opConst'.

Note that because this returns an Op, it can be used with any function that expects an OpM, as well.

opConst :: (Every Num as, Known Length as) => a -> Op as a

An Op that ignores all of its inputs and returns a given constant value.

>>> gradOp' (opConst 10) (1 ::< 2 ::< 3 ::< Ø)
(10, 0 ::< 0 ::< 0 ::< Ø)

opConst' :: Prod Summer as -> a -> Op as a

A version of opConst that takes explicit Summers, so can be run on values of types that aren't Num instances.

Giving gradients directly

op1' :: (a -> (b, Maybe b -> a)) -> Op '[a] b

Create an Op of a function taking one input, by giving its explicit derivative. The function should return a tuple containing the result of the function, and also a function taking the derivative of the result and return the derivative of the input.

If we have

\[ \eqalign{ f &: \mathbb{R} \rightarrow \mathbb{R}\cr y &= f(x)\cr z &= g(y) } \]

Then the derivative \( \frac{dz}{dx} \), it would be:

\[ \frac{dz}{dx} = \frac{dz}{dy} \frac{dy}{dx} \]

If our Op represents \(f\), then the second item in the resulting tuple should be a function that takes \(\frac{dz}{dy}\) and returns \(\frac{dz}{dx}\).

If the input is Nothing, then \(\frac{dz}{dy}\) should be taken to be \(1\).

As an example, here is an Op that squares its input:

square :: Num a => Op '[a] a
square = op1' $ x -> (x*x, case Nothing -> 2 * x
                                  Just d  -> 2 * d * x
                      )

Remember that, generally, end users shouldn't directly construct Ops; they should be provided by libraries or generated automatically.

For numeric functions, single-input Ops can be generated automatically using op1.

op2' :: (a -> b -> (c, Maybe c -> (a, b))) -> Op '[a, b] c

Create an Op of a function taking two inputs, by giving its explicit gradient. The function should return a tuple containing the result of the function, and also a function taking the derivative of the result and return the derivative of the input.

If we have

\[ \eqalign{ f &: \mathbb{R}^2 \rightarrow \mathbb{R}\cr z &= f(x, y)\cr k &= g(z) } \]

Then the gradient \( \left< \frac{\partial k}{\partial x}, \frac{\partial k}{\partial y} \right> \) would be:

\[ \left< \frac{\partial k}{\partial x}, \frac{\partial k}{\partial y} \right> = \left< \frac{dk}{dz} \frac{\partial z}{dx}, \frac{dk}{dz} \frac{\partial z}{dy} \right> \]

If our Op represents \(f\), then the second item in the resulting tuple should be a function that takes \(\frac{dk}{dz}\) and returns \( \left< \frac{\partial k}{dx}, \frac{\partial k}{dx} \right> \).

If the input is Nothing, then \(\frac{dk}{dz}\) should be taken to be \(1\).

As an example, here is an Op that multiplies its inputs:

mul :: Num a => Op '[a, a] a
mul = op2' $ x y -> (x*y, case Nothing -> (y  , x  )
                                 Just d  -> (d*y, x*d)
                     )

Remember that, generally, end users shouldn't directly construct Ops; they should be provided by libraries or generated automatically.

For numeric functions, two-input Ops can be generated automatically using op2.

op3' :: (a -> b -> c -> (d, Maybe d -> (a, b, c))) -> Op '[a, b, c] d

Create an Op of a function taking three inputs, by giving its explicit gradient. See documentation for op2' for more details.

Automatic creation using the ad library

op1 :: Num a => (forall s. AD s (Forward a) -> AD s (Forward a)) -> Op '[a] a

Automatically create an Op of a numerical function taking one argument. Uses diff, and so can take any numerical function polymorphic over the standard numeric types.

>>> gradOp' (op1 (recip . negate)) (5 ::< Ø)
(-0.2, 0.04 ::< Ø)

op2 :: Num a => (forall s. Reifies s Tape => Reverse s a -> Reverse s a -> Reverse s a) -> Op '[a, a] a

Automatically create an Op of a numerical function taking two arguments. Uses grad, and so can take any numerical function polymorphic over the standard numeric types.

>>> gradOp' (op2 (\x y -> x * sqrt y)) (3 ::< 4 ::< Ø)
(6.0, 2.0 ::< 0.75 ::< Ø)

op3 :: Num a => (forall s. Reifies s Tape => Reverse s a -> Reverse s a -> Reverse s a -> Reverse s a) -> Op '[a, a, a] a

Automatically create an Op of a numerical function taking three arguments. Uses grad, and so can take any numerical function polymorphic over the standard numeric types.

>>> gradOp' (op3 (\x y z -> (x * sqrt y)**z)) (3 ::< 4 ::< 2 ::< Ø)
(36.0, 24.0 ::< 9.0 ::< 64.503 ::< Ø)

opN :: (Num a, Known Nat n) => (forall s. Reifies s Tape => Vec n (Reverse s a) -> Reverse s a) -> Op (Replicate n a) a

Automatically create an Op of a numerical function taking multiple arguments. Uses grad, and so can take any numerical function polymorphic over the standard numeric types.

>>> gradOp' (opN (\(x :+ y :+ Ø) -> x * sqrt y)) (3 ::< 4 ::< Ø)
(6.0, 2.0 ::< 0.75 ::< Ø)

From Isomorphisms

opCoerce :: (Coercible a b, Num a) => Op '[a] b

An Op between two coercibles. Requires the input to be an instance of Num.

>>> gradOp' opCoerce (Identity 5) :: (Int, Identity Int)
(5, Identity 1)

opTup :: (Every Num as, Known Length as) => Op as (Tuple as)

An Op that takes as and returns exactly the input tuple.

>>> gradOp' opTup (1 ::< 2 ::< 3 ::< Ø)
(1 ::< 2 ::< 3 ::< Ø, 1 ::< 1 ::< 1 ::< Ø)

opIso :: Num a => Iso' a b -> Op '[a] b

An Op that runs the input value through the isomorphism encoded in the Iso. Requires the input to be an instance of Num.

Warning: This is unsafe! It assumes that the isomorphisms themselves have derivative 1, so will break for things like exponentiating. Basically, don't use this for any "numeric" isomorphisms.

opCoerce' :: Coercible a b => Unity a -> Op '[a] b

A version of opCoerce that takes an explicit Unity, so can be run on values that aren't Num instances.

opTup' :: Prod Unity as -> Op as (Tuple as)

A version of opTup that takes explicit Unitys, so can be run on values of types that aren't Num instances.

opIso' :: Unity a -> Iso' a b -> Op '[a] b

A version of opIso that takes an explicit Unity, so can be run on values of types that aren't Num instances.

Utility

type family Replicate (n :: N) (a :: k) = (as :: [k]) | as -> n where ...

Equations

Replicate Z a = '[]

Replicate (S n) a = a ': Replicate n a