Numeric.Backprop.Op.Mono

Description

Provides monomorphic versions of the types and combinators in Numeric.Backprop.Op, for usage with Numeric.Backprop.Mono and Numeric.Backprop.Mono.Implicit.

Note that Op is a subset or subtype of OpM, and so, any function that expects an OpM m as a (or an OpB s as a) can be given an Op as a and it'll work just fine.

Types

type Op n a b = Op (Replicate n a) b

type OpM m n a = OpM m (Replicate n a)

data VecT k n f a :: forall k. N -> (k -> *) -> k -> * where

Constructors

ØV :: VecT k Z f a

(:*) :: VecT k (S n1) f a infixr 4

type Vec n = VecT * n I

Running

runOp' :: Op n a b -> Vec n a -> (b, Maybe b -> Vec n a)

runOp :: Op n a b -> Vec n a -> b

gradOp :: Op n a b -> Vec n a -> Vec n a

gradOpWith :: Op n a b -> Vec n a -> b -> Vec n a

gradOpWith' :: Op n a b -> Vec n a -> Maybe b -> Vec n a

Creation

op1 :: Num a => (forall s. AD s (Forward a) -> AD s (Forward a)) -> Op N1 a a

op2 :: Num a => (forall s. Reifies s Tape => Reverse s a -> Reverse s a -> Reverse s a) -> Op N2 a a

op3 :: Num a => (forall s. Reifies s Tape => Reverse s a -> Reverse s a -> Reverse s a -> Reverse s a) -> Op N3 a a

opN :: (Num a, Known Nat n) => (forall s. Reifies s Tape => Vec n (Reverse s a) -> Reverse s a) -> Op n a a

Utility

Vectors

pattern (:+) :: forall a n. a -> Vec n a -> Vec (S n) a infixr 4

(*:) :: f a -> f a -> VecT k (S (S Z)) f a infix 5

(+:) :: a -> a -> Vec (S (S Z)) a infix 5

head' :: VecT k (S n) f a -> f a

Type synonyms

type N0 = Z

Convenient aliases for low-value Peano numbers.

type N1 = S N0

type N2 = S N1

type N3 = S N2

type N4 = S N3

type N5 = S N4

type N6 = S N5

type N7 = S N6

type N8 = S N7

type N9 = S N8

type N10 = S N9