Numeric.Backprop.Tuple

Description

Canonical strict tuples (and unit) with Num instances for usage with backprop. This is here to solve the problem of orphan instances in libraries and potential mismatched tuple types.

If you are writing a library that needs to export BVars of tuples, consider using the tuples in this module so that your library can have easy interoperability with other libraries using backprop.

Because of API decisions, backprop and gradBP only work with things with Num instances. However, this disallows default Prelude tuples (without orphan instances from packages like NumInstances).

Until tuples have Num instances in base, this module is intended to be a workaround for situations where:

This comes up often in cases where:

A function wants to return more than one value (BVar s (T2 a b)

You want to uncurry a BVar function to use with backprop and gradBP.

You want to use the useful Prisms automatically generated by the lens library, which use tuples for multiple-constructor fields.

Only 2-tuples and 3-tuples are provided. Any more and you should probably be using your own custom product types, with instances automatically generated from something like one-liner-instances.

Lenses into the fields are provided, but they also work with _1, _2, and _3 from Lens.Micro. However, note that these are incompatible with _1, _2, and _3 from Control.Lens.

You can "construct" a BVar s (T2 a b) with functions like isoVar.

Since: 0.1.1.0

Zero-tuples (unit)

data T0

Unit ('()') with Num, Fractional, and Floating instances.

Be aware that the methods in its numerical instances are all non-strict:

@ _ + _ = T0 negate _ = T0 fromIntegral _ = T0 @

Since: 0.1.4.0

Constructors

T0

Two-tuples

data T2 a b

Strict 2-tuple with Num, Fractional, and Floating instances.

Since: 0.1.1.0

Constructors

T2 !a !b

Conversions

If using lens, the two conversion functions can be chained with prisms and traversals and other optics using:

iso tupT2 t2Tup :: Iso' (a, b) (T2 a b)

t2Tup :: T2 a b -> (a, b)

Convert to a Haskell tuple.

Forms an isomorphism with tupT2.

tupT2 :: (a, b) -> T2 a b

Convert from Haskell tuple.

Forms an isomorphism with t2Tup.

Consumption

uncurryT2 :: (a -> b -> c) -> T2 a b -> c

Uncurry a function to take in a T2 of its arguments

Since: 0.1.2.0

curryT2 :: (T2 a b -> c) -> a -> b -> c

Curry a function taking a T2 of its arguments

Since: 0.1.2.0

Lenses

t2_1 :: Lens (T2 a b) (T2 a' b) a a'

Lens into the first field of a T2. Also exported as _1 from Lens.Micro.

t2_2 :: Lens (T2 a b) (T2 a b') b b'

Lens into the second field of a T2. Also exported as _2 from Lens.Micro.

Three-tuples

data T3 a b c

Strict 3-tuple with a Num, Fractional, and Floating instances.

Since: 0.1.1.0

Constructors

T3 !a !b !c

Conversions

If using lens, the two conversion functions can be chained with prisms and traversals and other optics using:

iso tupT3 t2Tup :: Iso' (a, b, c) (T3 a b c)

t3Tup :: T3 a b c -> (a, b, c)

Convert to a Haskell tuple.

Forms an isomorphism with tupT3.

tupT3 :: (a, b, c) -> T3 a b c

Convert from Haskell tuple.

Forms an isomorphism with t3Tup.

Lenses

t3_1 :: Lens (T3 a b c) (T3 a' b c) a a'

Lens into the first field of a T3. Also exported as _1 from Lens.Micro.

t3_2 :: Lens (T3 a b c) (T3 a b' c) b b'

Lens into the second field of a T3. Also exported as _2 from Lens.Micro.

t3_3 :: Lens (T3 a b c) (T3 a b c') c c'

Lens into the third field of a T3. Also exported as _3 from Lens.Micro.

Consumption

uncurryT3 :: (a -> b -> c -> d) -> T3 a b c -> d

Uncurry a function to take in a T3 of its arguments

Since: 0.1.2.0

curryT3 :: (T3 a b c -> d) -> a -> b -> c -> d

Curry a function taking a T3 of its arguments

Since: 0.1.2.0